Improved UI, enemies and Skills
Now the UI shows the face of the main character and it changes depending on his current life. The enemies had suffered a collider minimal change and added that they change the order in layer depending on their position and the position of the player. 
Finally also added two more skills.l