Improved the UI, Skill Tree and fixed the MiniMap
The minimap doesn't renders the particles anymore, the UI shows the buttons you can use to play, the Skill Tree shows the points needed to unlock the abilities and the abilities have their own icon. Finally when you are near to an object a pop-up shows in the screen showing the player the button he has to push to interact with it.